refactor(Message): Message 클래스 view 패키지로 분리
