+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import constant.Message;
+import view.Message;
 
 public class Application {
     public static void main(String[] args) {
